tests/wpt/web-platform-tests/app-history/navigate-event/navigateerror-ordering-location-api-reentrant.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  let start_href = location.href;
  let events = [];
  let first_onnavigate = true;
  appHistory.onnavigatesuccess = () => events.push("navigatesuccess");
  appHistory.onnavigateerror = t.step_func(() => {
    events.push("navigateerror");
    assert_equals(location.hash, "");
  });
  appHistory.onnavigate = t.step_func(e => {
    events.push("navigate");
    e.signal.onabort = () => events.push("abort");
    e.transitionWhile(Promise.resolve());
    if (first_onnavigate) {
      first_onnavigate = false;
      location.hash = "#2";
    }
  });
  location.hash = "#1";
  location.hash = "#2";
  await new Promise(resolve => t.step_timeout(resolve, 0));
  assert_equals(location.hash, "#2");
  assert_array_equals(events, ["navigate", "abort", "navigateerror", "navigate", "navigatesuccess"]);
}, "navigateerror ordering when navigating via the location API reentrantly");
</script>
